requirements.txt
Flask==2.2.2
twilio==7.7.0
python-dotenv==0.19.1
openai==0.25.0
heyoo==0.0.7
pyngrok==5.2.1
gunicorn==19.9.0